# Conflicts:
#	gradle/libs.versions.toml
#	localDatasource/build.gradle.kts
#	localDatasource/src/main/java/com/example/localdatasource/roomDataBase/datasource/MovieLocalDataSourceImpl.kt
#	localDatasource/src/main/java/com/example/localdatasource/roomDataBase/datasource/TvShowLocalDataSourceImpl.kt
#	localDatasource/src/test/java/com/example/localdatasource/TvShowLocalDataSourceImplTest.kt

+++ repository/src/main/java/com/example/repository/mapper/remoteToLocal/TvShowCategoryRemoteLocalMapper.kt
@@ -8,7 +8,7 @@ class TvShowCategoryRemoteLocalMapper :
     RemoteToLocalMapper<RemoteCategoryDto, LocalTvShowCategoryDto> {
     override fun toLocal(remote: RemoteCategoryDto, args: List<Any>): LocalTvShowCategoryDto {
         return LocalTvShowCategoryDto(
-            categoryId = remote.id,
+            categoryId = remote.id.toLong(),
             storedLanguage = args.first().toString(),
             name = remote.name
         )
